Conflate 3 cpp files to shrink the package size
